+++ Toggly.FeatureManagement.Remix/remix-toggly-server/tsup.config.ts
@@ -9,6 +9,9 @@ export default defineConfig({
   target: 'es2020',
   external: [
     '@ops-ai/remix-toggly-core',
+    '@ops-ai/remix-toggly-core/telemetry/grpc',
+    '@grpc/grpc-js',
+    '@grpc/proto-loader',
     '@remix-run/node',
     '@remix-run/server-runtime',
   ],